Committed changes

+++ src/sk/com/ymca/javaStart/lecture8/classwork/ex1/ArithmeticOperations.java
@@ -1,19 +1,19 @@
 package sk.com.ymca.javaStart.lecture8.classwork.ex1;
 
 public class ArithmeticOperations {
-    public double add(double number1, double number2) {
+    public double add(int number1, int number2) {
         return number1 + number2;
     }
 
-    public double sub(double number1, double number2) {
+    public double sub(int number1, int number2) {
         return number1 - number2;
     }
 
-    public double mul(double number1, double number2) {
+    public double mul(int number1, int number2) {
         return number1 * number2;
     }
 
-    public double div(double number1, double number2) {
+    public double div(int number1, int number2) {
         return number1 / number2;
     }
 }
